+++ Dolphin/Core/Application.h
@@ -1,26 +1,3 @@
 ﻿#pragma once
-#include "DolphinMacro.h"
-#include "DolphinStd.h"
-#include "ObjectBuilder.h"
-
-
-namespace Dolphin
-{
-	namespace Core
-	{
-		class Application
-		{
-		private:
-			Object* rootObject;
-			bool quit;
-
-
-		public:
-			Application();
-			~Application();
-			void Tick();
-			void Quit();
-			bool IsQuit();
-		};
-	}
-}
+#include "Application/Application_forward.h"
+#include "Application/Application_member.h"
