+++ Dolphin/Core/Application/Application_forward.h
@@ -1,0 +1,10 @@
+#pragma once
+
+
+namespace Dolphin
+{
+	namespace Core
+	{
+		class Application;
+	}
+}
