--- Dolphin/Core/Component.cpp
@@ -1,8 +1,0 @@
-﻿#include "Component.h"
-
-DolphinCore::Component::Component(Object* object)
-{
-	this->object = object;
-}
-
-DolphinCore::Object* DolphinCore::Component::ThisObject() { return this->object; }
