+++ Dolphin/Core/Component.h
@@ -1,28 +1,3 @@
 ﻿#pragma once
-#include "DolphinMacro.h"
-#include "DolphinStd.h"
-
-
-namespace Dolphin
-{
-	namespace Core
-	{
-		class Object;
-		class Component
-		{
-		friend class Object;
-
-		protected:
-			Object* object;
-
-			Component(Object* object);
-			virtual ~Component(){};
-			virtual void Start(){};
-			virtual void Tick(){};
-
-
-		public:
-			Object* ThisObject();
-		};
-	}
-}
+#include "Component/Component_forward.h"
+#include "Component/Component_member.h"
