+++ Dolphin/Core/Component/Component.cpp
@@ -1,0 +1,8 @@
+﻿#include "Component.h"
+
+Dolphin::Core::Component::Component(Object* object)
+{
+	this->object = object;
+}
+
+Dolphin::Core::Object* Dolphin::Core::Component::ThisObject() { return this->object; }
