--- Dolphin/Core/Component/Nest.h
@@ -1,31 +1,0 @@
-#pragma once
-#include "DolphinMacro.h"
-#include "DolphinStd.h"
-#include "../Component.h"
-#include "../Object.h"
-
-
-namespace Dolphin
-{
-	namespace StandardComponent
-	{
-		class Object;
-		class Nest : public DolphinCore::Component
-		{
-		private:
-			DolphinCore::Object* parent;
-			vector<DolphinCore::Object*> children;
-
-
-		public:
-			Nest(DolphinCore::Object* object);
-			~Nest() override;
-			int ChildCount();
-			DolphinCore::Object* Parent();
-			DolphinCore::Object* MoveTo(DolphinCore::Object* target);
-			DolphinCore::Object* GetChild(string name);
-
-			void Start() override {};
-		};
-	}
-}
