+++ Dolphin/Core/ObjectBuilder/BuilderChain/BuilderChain.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "BuilderChain_forward.h"
+#include "BuilderChain_member.h"
+#include "BuilderChain_inline.h"
