+++ Dolphin/Core/ObjectBuilder/BuilderChain/BuilderChain_forward.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace Dolphin
+{
+	class BuilderChain;
+}
