+++ Dolphin/Core/ObjectBuilder/ObjectBuilder_forward.h
@@ -1,0 +1,7 @@
+#pragma once
+
+
+namespace Dolphin
+{
+	class ObjectBuilder;
+}
