+++ Dolphin/Core/ObjectBuilder/ObjectBuilder_member.h
@@ -1,0 +1,15 @@
+#pragma once
+#include "DolphinStd.h"
+#include "Object/Object_member.h"
+#include "ObjectBuilder_forward.h"
+#include "BuilderChain/BuilderChain.h"
+
+
+namespace Dolphin
+{
+	class ObjectBuilder
+	{
+	public:
+		static BuilderChain Instantiate(string name);
+	};
+}
