+++ Dolphin/Header/DolphinMacro.h
@@ -1,9 +1,5 @@
 #pragma once
 
-// 名前空間のエイリアス
-#define DolphinCore Dolphin::Core
-#define DolphinSC Dolphin::StandardComponent
-
 
 // 便利マクロ
 #define FOR(index, maxCount) for(int index = 0; index < maxCount; ++index)
